Wrist-rest polyhedrons.
Unified and extended the concept of named anchors, adding a :derived
section for these to the top level of getopt, and a reckon-feature
function that acts as a clearing house.

Separated user parameter parsing from validation, overhauling the
relationship between the parameter modules.

Major changes to wrist-rest-related parameters and possibilities.

Prettier user-facing exceptions on configuration errors.

Replaced lots of instances of vec on map with mapv.

Fixed a bug in no_bottom.yaml. Added a corresponding configuration
fragment for a back plate.